Refactor: 로그인 시 토큰이 바디로 반환되도록 수정
