+++ src/main/java/sportsmatchingservice/auth/filter/AuthenticationFilter.java
@@ -7,30 +7,19 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import sportsmatchingservice.auth.dto.LoginDto;
-import sportsmatchingservice.auth.jwt.JwtTokenizer;
 import sportsmatchingservice.constant.ErrorCode;
 import sportsmatchingservice.exceptions.exception.GeneralException;
-import sportsmatchingservice.auth.domain.User;
 
 import javax.servlet.FilterChain;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 
 
 public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
     private final AuthenticationManager authenticationManager;
-    private final JwtTokenizer jwtTokenizer;
 
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-
-    public AuthenticationFilter(AuthenticationManager authenticationManager, JwtTokenizer jwtTokenizer) {
+    public AuthenticationFilter(AuthenticationManager authenticationManager) {
         this.authenticationManager = authenticationManager;
-        this.jwtTokenizer = jwtTokenizer;
     }
 
     @SneakyThrows
@@ -49,47 +38,12 @@ public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
     protected void successfulAuthentication(HttpServletRequest request,
                                             HttpServletResponse response,
                                             FilterChain chain,
-                                            Authentication authResult) throws IOException {
+                                            Authentication authResult) {
         try {
-            User user = (User) authResult.getPrincipal();
-
-            String accessToken = delegateAccessToken(user);
-            String refreshToken = delegateRefreshToken(user);
-
-            response.setHeader("Authorization", "Bearer" + accessToken);
-            response.setHeader("Refresh", refreshToken);
-
             this.getSuccessHandler().onAuthenticationSuccess(request, response, authResult);
 
         } catch (Exception e) {
             throw new GeneralException(ErrorCode.INTERNAL_ERROR, e);
         }
-    }
-
-
-    private String delegateAccessToken(User user) {
-        Map<String, Object> claims = new HashMap<>();
-        claims.put("username", user.getEmail());
-        claims.put("roles", user.getRoles());
-
-        String subject = user.getEmail();
-        Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
-
-        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-
-        String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
-
-        return accessToken;
-    }
-
-    private String delegateRefreshToken(User user) {
-        String subject = user.getEmail();
-        Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getRefreshTokenExpirationMinutes());
-
-        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-
-        String refreshToken = jwtTokenizer.generateRefreshToken(subject, expiration, base64EncodedSecretKey);
-
-        return refreshToken;
     }
 }
